Moved ChessMove to service package
Since it provides a  request scoped service.
